requirements.txt
moviepy==1.0.3
Pillow==9.5.0
edge-tts==6.1.9
requests==2.31.0
numpy==1.24.3
